requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.1
django-admin-bootstrapped==2.5.0
django-bootstrap3==5.4.0
django-crispy-forms==1.4.0
django-datetime-widget==0.9.3
django-easy-maps==0.9
django-toolbelt==0.0.1
geopy==1.10.0
gunicorn==19.3.0
psycopg2==2.6
pytz==2015.2
static3==0.5.1
